+++ xiuwen-backend-framework/xiuwen-web/src/main/java/com/xiuwen/web/controller/merchant/MerchantCustomDesignController.java
@@ -45,8 +45,8 @@ public class MerchantCustomDesignController {
 
         return Result.success(PageResult.of(
                 pageResult.getTotal(),
-                (int) pageResult.getCurrent(),
-                (int) pageResult.getSize(),
+                pageResult.getCurrent(),
+                pageResult.getSize(),
                 voList));
     }
