+++ xiuwen-backend-framework/xiuwen-web/src/main/java/com/xiuwen/web/controller/merchant/MerchantOrderController.java
@@ -65,8 +65,8 @@ public class MerchantOrderController {
 
         return Result.success(PageResult.of(
                 pageResult.getTotal(),
-                (int) pageResult.getCurrent(),
-                (int) pageResult.getSize(),
+                pageResult.getCurrent(),
+                pageResult.getSize(),
                 voList
         ));
     }
