+++ xiuwen-backend-framework/xiuwen-web/src/main/java/com/xiuwen/web/controller/merchant/MerchantPatternController.java
@@ -56,8 +56,8 @@ public class MerchantPatternController {
 
         return Result.success(PageResult.of(
                 page.getTotal(),
-                (int) page.getCurrent(),
-                (int) page.getSize(),
+                page.getCurrent(),
+                page.getSize(),
                 voList
         ));
     }
